requirements.txt
livekit-agents[images]
livekit-plugins-google
livekit-plugins-silero
livekit-plugins-turn-detector
python-dotenv
